chore: Bridge 인스턴스를 사용하여 핸드쉐이크 및 요청에 대한 응답을 핸들링할 수 있도록 구현
#5
